requirements.txt

fastapi
uvicorn
pydantic[dotenv]
redis
python-multipart
passlib[bcrypt]
python-jose
sqlalchemy
databases